Move get_or_create_reserve_id to utils, match reserve_id in AJAX call to reserve asset
