+++ concordia/templates/transcriptions/asset_detail.html
@@ -554,11 +554,11 @@
 
 <script>
     {% if transcription_status == "not_started" or transcription_status == "in_progress" %}
-        attemptToReserveAsset("{% url 'reserve-asset' asset.pk %}",
+        attemptToReserveAsset("{% url 'reserve-asset' asset.pk reserve_id %}",
         "",
         "transcribe", true);
     {% elif user.is_authenticated %}
-        attemptToReserveAsset("{% url 'reserve-asset' asset.pk %}",
+        attemptToReserveAsset("{% url 'reserve-asset' asset.pk reserve_id %}",
         "{% url 'transcriptions:redirect-to-next-reviewable-asset' asset.item.project.campaign.slug %}",
         "review", true);
     {% endif %}
